requirements.txt
customtkinter==5.2.2
darkdetect==0.8.0
et-xmlfile==1.1.0
lxml==5.2.2
numpy==2.0.0
openpyxl==3.1.4
packaging==24.1
pandas==2.2.2
python-dateutil==2.9.0.post0
python-docx==1.1.2
pytz==2024.1
six==1.16.0
typing_extensions==4.12.2
tzdata==2024.1
